新增销售单导出excel功能

+++ src/main/java/com/mf/util/ExcelUtil.java
@@ -120,7 +120,7 @@ public class ExcelUtil {
 
         // 创建列首-增加样式-赋值
         HSSFRow row = sheet.createRow(1);
-        for (short i = 0; i < headers.length; i++) {
+        for (int i = 0; i < headers.length; i++) {
 
             @SuppressWarnings("deprecation")
             HSSFCell cell = row.createCell(i);
@@ -140,7 +140,7 @@ public class ExcelUtil {
             // 利用反射，根据javabean属性的先后顺序，动态调用getXxx()方法得到属性值
             Field[] fields = t.getClass().getDeclaredFields();
             linenum = index - 1;
-            for (short i = 0; i < fields.length; i++) {
+            for (int i = 0; i < fields.length; i++) {
                 @SuppressWarnings("deprecation")
                 HSSFCell cell = row.createCell(i);
                 cell.setCellStyle(dataSetStyle);
